refactor(ui): canonicalize totp verify module filename
